+++ lib/emailbutler/engine.rb
@@ -3,5 +3,11 @@
 module Emailbutler
   class Engine < ::Rails::Engine
     isolate_namespace Emailbutler
+
+    config.generators do |g|
+      g.test_framework :rspec
+      g.fixture_replacement :factory_bot
+      g.factory_bot dir: 'spec/factories'
+    end
   end
 end