+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - saving additional mail params to message
+- Webhooks::Receiver class for checking webhooks payload and updating messages
 
 ## [0.1.0] - 2022-09-01
 ### Added